Remove merge artifacts v1
