--- ecdk/src/command/__init__.py
@@ -1,14 +1,0 @@
-"""
-Implementations of particular commands
-"""
-
-from .run import run
-from .analyze import analyze
-from .perfcmp import perfcmp
-
-__all__ = (
-    run,
-    analyze,
-    perfcmp
-)
-
