+++ ecdk/src/ecdk/command/perfcmp.py
@@ -1,7 +1,7 @@
-from core.env import EnvContext
 from cli.args import PerfcmpCmdArgs
 from data.processing import compare_exp_batch_outputs
+from context import Context
 
 
-def perfcmp(ctx: EnvContext, args: PerfcmpCmdArgs):
+def perfcmp(ctx: Context, args: PerfcmpCmdArgs):
     compare_exp_batch_outputs(args.basepath, args.benchpath)
